Update nutrition_code.ipynb
began cleaning the data, consolidating the dataset to the most important columns for my goals and checked those columns for uniformity in units.

Next step is to strip the units, add those abbreviations into the column names themselves to maintain their information, then deal with the NaN values in saturated fats.