Calculator Attempts
So for about 4 days I've been trying to create a nutrition calculator/visualization that will take user input ingredients and serving size then produce a table of the complied ingredient values and charts of the total nutrient counts.

The nutrition_calculator_one.ipynb is the best success but still needs to be cleaned and organized.